# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Threat Modeling Agent API
  version: 0.1.0
  description: |
    REST API for the Threat Modeling Agent SaaS platform.

    ## Authentication
    All endpoints (except the auth callback) require a valid Bearer token issued by WorkOS.
    Tokens are validated server-side on every request. Client-side validation is not sufficient.

    ## Tenant scoping
    All data is scoped to an organisation. The `org_id` is extracted from the validated JWT —
    it is never accepted from request parameters or body.

    ## Rate limiting
    Authentication and job submission endpoints are rate-limited.
    See individual endpoint descriptions for details.

    ## Security
    This API is governed by CLAUDE.md and docs/specs/06-security.md.
    All responses to authenticated requests include `Cache-Control: no-store`.
servers:
  - url: https://api.tmsa.example.com/v1
    description: Production (West Europe)
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: WorkOS-issued JWT. Validated server-side on every request.
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Organisation UUID. Must match the org_id in the authenticated user's JWT.
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ThreatId:
      name: threatId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    PageCursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque pagination cursor from a previous response.
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code. Never contains stack traces or internal paths.
        message:
          type: string
          description: Human-readable summary. Never contains internal details.
        correlationId:
          type: string
          format: uuid

    # ── Pagination ───────────────────────────────────────────────────────────────
    PaginationMeta:
      type: object
      properties:
        nextCursor:
          type: string
          nullable: true
        hasMore:
          type: boolean
        totalCount:
          type: integer

    # ── Auth ─────────────────────────────────────────────────────────────────────
...
    JobStatus:
      type: string
      enum:
        - PENDING
        - PARSING
        - NORMALIZING
        - AWAITING_REVIEW
        - CLASSIFYING
        - ANALYZING
        - SYNTHESIZING
        - COMPLETE
        - FAILED
        - PARTIAL
    JobSummary:
      type: object
      required: [id, title, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/JobStatus'
        artifactType:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
    JobDetail:
      allOf:
        - $ref: '#/components/schemas/JobSummary'
        - type: object
          properties:
            errorCode:
              type: string
              nullable: true
            threatCount:
              type: integer
              nullable: true
            confirmedThreatCount:
              type: integer
              nullable: true

    # ── Architecture ─────────────────────────────────────────────────────────────
...
    Threat:
      type: object
      required: [id, identifier, title, methodCategory, findingType, status, confidence, evidenceStrength, source]
      properties:
        id:
          type: string
          format: uuid
        identifier:
          type: string
          example: T-001
        title:
          type: string
        methodCategory:
          type: string
        affectedElementIds:
          type: array
          items:
            type: string
            format: uuid
        description:
          type: string
        attackScenario:
          type: string
        preconditions:
          type: string
          nullable: true
        impactedAssets:
          type: array
          items:
            type: string
        securityImpact:
          type: string
          nullable: true
        privacyImpact:
          type: string
          nullable: true
        existingControls:
          type: string
          nullable: true
        controlGaps:
          type: string
          nullable: true
        confidence:
          type: string
          enum: [high, medium, low]
        evidenceBasis:
          type: array
          items:
            type: string
        evidenceStrength:
          type: string
          enum: [direct, inferred, assumption_dependent]
        assumptions:
          type: string
          nullable: true
        findingType:
          type: string
          enum: [confirmed, conditional, user_added]
        status:
          type: string
          enum: [open, accepted, mitigated, rejected]
        source:
          type: string
          enum: [system, user]
        mitigations:
          type: array
          items:
            $ref: '#/components/schemas/Mitigation'
        frameworkMappings:
          type: array
          items:
            $ref: '#/components/schemas/FrameworkMapping'
        notes:
          type: array
          items:
            $ref: '#/components/schemas/ThreatNote'
    Mitigation:
      type: object
      required: [id, title, description, priority]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [critical, high, medium, low]
        category:
          type: string
          nullable: true
    FrameworkMapping:
      type: object
      required: [framework, reference, mappingType]
      properties:
        framework:
          type: string
          enum: [owasp_top10, owasp_api_top10, asvs, cis_controls, ncsc, twelve_factor]
        reference:
          type: string
        mappingType:
          type: string
          enum: [direct, approximate]
    ThreatNote:
      type: object
      required: [id, body, createdBy, createdAt]
      properties:
        id:
          type: string
          format: uuid
        body:
          type: string
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    AddThreatRequest:
      type: object
      required: [title, methodCategory, description, attackScenario, affectedElementIds]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        methodCategory:
          type: string
          maxLength: 100
        affectedElementIds:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
        description:
          type: string
          minLength: 1
          maxLength: 10000
        attackScenario:
          type: string
          minLength: 1
          maxLength: 10000
        preconditions:
          type: string
          maxLength: 5000
        impactedAssets:
          type: array
          items:
            type: string
            maxLength: 255
        securityImpact:
          type: string
          maxLength: 5000
        privacyImpact:
          type: string
          maxLength: 5000
    UpdateThreatStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [open, accepted, mitigated, rejected]
        note:
          type: string
          maxLength: 5000
    AddThreatNoteRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 10000

    # ── Analysis output ──────────────────────────────────────────────────────────
    AnalysisSummary:
      type: object
      required: [jobId, status, confirmedThreats, conditionalThreats, userAddedThreats]
      properties:
        jobId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/JobStatus'
        systemSummary:
          type: string
          nullable: true
        architectureClassification:
          type: array
          items:
            type: string
        selectedMethods:
          type: array
          items:
            $ref: '#/components/schemas/SelectedMethod'
        modelRoutingSummary:
          type: object
          additionalProperties:
            type: string
        confirmedThreats:
          type: integer
        conditionalThreats:
          type: integer
        userAddedThreats:
          type: integer
        reviewQuestions:
          type: array
          items:
            type: string
    SelectedMethod:
      type: object
      required: [method, rationale]
      properties:
        method:
          type: string
        rationale:
          type: string

# ── Paths ──────────────────────────────────────────────────────────────────────
paths:
...
  /orgs/{orgId}/jobs/{jobId}:
    parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/JobId'

    get:
      operationId: getJob
      summary: Get job status and summary
      tags: [jobs]
      responses:
        '200':
          description: Job detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDetail'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: deleteJob
      summary: Delete a job and all its associated data
      tags: [jobs]
      responses:
        '204':
          description: Job deleted
        '409':
          description: Job is in progress and cannot be deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Architecture (canonical model) ───────────────────────────────────────────
...
  /orgs/{orgId}/jobs/{jobId}/analysis:
    parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/JobId'

    get:
      operationId: getAnalysis
      summary: Get the full analysis summary for a completed job
      description: Available once job status is COMPLETE or PARTIAL.
      tags: [analysis]
      responses:
        '200':
          description: Analysis summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisSummary'
        '409':
          description: Analysis not yet complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Threats ──────────────────────────────────────────────────────────────────
  /orgs/{orgId}/jobs/{jobId}/threats:
    parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/JobId'

    get:
      operationId: listThreats
      summary: List threats for a job
      tags: [threats]
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PageCursor'
        - name: findingType
          in: query
          schema:
            type: string
            enum: [confirmed, conditional, user_added]
        - name: status
          in: query
          schema:
            type: string
            enum: [open, accepted, mitigated, rejected]
        - name: elementId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter threats affecting a specific architecture element
      responses:
        '200':
          description: Threat list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Threat'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      operationId: addThreat
      summary: Add a user-defined threat to the analysis
      tags: [threats]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddThreatRequest'
      responses:
        '201':
          description: Threat added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Threat'
  /orgs/{orgId}/jobs/{jobId}/threats/{threatId}:
    parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/JobId'
      - $ref: '#/components/parameters/ThreatId'

    get:
      operationId: getThreat
      summary: Get a single threat with full detail
      tags: [threats]
      responses:
        '200':
          description: Threat detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Threat'
        '404':
          description: Threat not found or does not belong to this job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orgs/{orgId}/jobs/{jobId}/threats/{threatId}/status:
    parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/JobId'
      - $ref: '#/components/parameters/ThreatId'

    patch:
      operationId: updateThreatStatus
      summary: Update a threat's status (accept, mitigate, reject, re-open)
      description: |
        Valid statuses: Open, Accepted, Mitigated, Rejected.
        Case-sensitive. The status field must match exactly.
      tags: [threats]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateThreatStatusRequest'
      responses:
        '200':
          description: Updated threat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Threat'
        '400':
          description: Invalid status value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orgs/{orgId}/jobs/{jobId}/export:
    parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/JobId'

    get:
      operationId: exportAnalysis
      summary: Download the full analysis output as a JSON file
      description: |
        Available once job status is COMPLETE or PARTIAL.
        Returns the raw analysis blob as a file download (Content-Disposition: attachment).
        Use GET /analysis for in-page rendering of the parsed data.
        Response includes Cache-Control: no-store.
      tags: [analysis]
      responses:
        '200':
          description: Analysis JSON file download
          content:
            application/json:
              schema:
                type: object
                description: Raw analysis output blob
          headers:
            Content-Disposition:
              schema:
                type: string
                example: 'attachment; filename="threat-model-{jobId}.json"'
        '404':
          description: Analysis output not found (job may not have completed successfully)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Analysis not yet complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orgs/{orgId}/jobs/{jobId}/threats/{threatId}/notes:
    parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/JobId'
      - $ref: '#/components/parameters/ThreatId'

    post:
      operationId: addThreatNote
      summary: Add a note to a threat
      tags: [threats]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddThreatNoteRequest'
      responses:
        '201':
          description: Note added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreatNote'
...